Refactored targets system. Merged static part of Target class with TargetsFolderSelector into Targets class.
