+++ Assets/Scripts/MultipleInstancesException.cs
@@ -1,0 +1,5 @@
+﻿class MultipleInstancesException : System.Exception
+{
+    public MultipleInstancesException()
+        : base("This class should have only one instance on the scene.") { }
+}
