+++ Assets/Scripts/MultipleInstancesException.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 494be9c69fb47994abdd57a36a16afb2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
