--- Assets/Scripts/TargetAnnotation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d05ff72398206ba4f8f349a4ca7215d7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
