+++ Assets/Scripts/Targets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0dce00d7a40d3be44bf26f571c06902d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
