+++ Assets/Scripts/Targets/Target.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Target : MonoBehaviour
+{
+    public CameraType cameraType;
+    public GameObject annotation;
+
+    private void Start()
+    {
+        if (annotation == null)
+        {
+            annotation = gameObject;
+        }
+    }
+}
