+++ Assets/Scripts/Targets/Targets.cs
@@ -1,0 +1,76 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[RequireComponent(typeof(ScenesGenerator))]
+public class Targets : MonoBehaviour
+{
+    public GameObject[] targetsFolders;
+    public int selected;
+    public bool useSettings = true;
+    
+    // these are static to simplify the static methods
+    [ResetParameter] static bool collectData = false;
+    [ResetParameter] static int focusedObject = 0;
+    [ResetParameter("Positive")] static bool positiveExamples = false;
+    static List<Target> targets=new List<Target>();
+    static Targets instance=null;
+
+    private void Awake()
+    {
+        if (instance != null)
+        {
+            throw new MultipleInstancesException();
+        }
+        instance = this;
+
+        ResetParameterAttribute.InitializeAll(this);
+        if (InitializedSettings.IsMenu)
+            selected = InitializedSettings.targetsFolderIndex;
+        SetFoldersActivation();
+
+        Transform folder = targetsFolders[selected].transform;
+        GetComponent<ScenesGenerator>().targetsFolder = folder;
+        ListTargets(folder);
+    }
+
+    void SetFoldersActivation()
+    {
+        for (int i = 0; i < targetsFolders.Length; i++)
+        {
+            targetsFolders[i].SetActive(i == selected);
+        }
+    }
+
+    static void ListTargets(Transform folder)
+    {
+        targets.Clear();
+        // it is important to obtain targets in their order in the scene hierarchy
+        // there is no such guarantee in FinObjectsOfType function
+        for (int i = 0; i < folder.childCount; i++)
+        {
+            Transform child = folder.GetChild(i);
+            if (child.gameObject.activeSelf)
+            {
+                Target target = child.GetComponent<Target>();
+                if (target != null)
+                    targets.Add(target);
+            }
+        }
+    }
+
+    public static Target Focused
+    {
+        get
+        {
+             if (collectData && !positiveExamples)
+                 return null;
+             if (focusedObject >= 0 && focusedObject < targets.Count)
+                 return targets[focusedObject];
+             else
+                 return null;
+        }
+    }
+
+    public static int Count=> targets.Count;
+}
